Inline Fedify execution context
