--- packages/fedify/src/fedify-execution.ts
@@ -1,10 +1,0 @@
-import { db } from '@kosmo/core/db';
-import type { DatabaseHandle } from '@kosmo/core/db';
-
-export type FedifyExecutionContext = {
-  readonly db: DatabaseHandle;
-};
-
-export const createFedifyExecutionContext = (
-  handle: DatabaseHandle = db,
-): FedifyExecutionContext => ({ db: handle });
