+++ packages/fedify/src/follow-delivery.test.ts
@@ -4,7 +4,7 @@ import { Accept, Follow, Person } from '@fedify/vocab';
 import { sendAcceptFollowActivity } from './follow-delivery';
 import type { Context } from '@fedify/fedify';
 import type { Activity, Recipient } from '@fedify/vocab';
-import type { FedifyExecutionContext } from './fedify-execution';
+import type { FedifyExecutionContext } from './federation';
 
 const canonicalOrigin = 'https://kos.moe';
 const senderProfileId = '019f6f67-1111-7777-8888-123456789abc';
