+++ packages/fedify/src/inbound-create.test.ts
@@ -35,14 +35,13 @@ import {
   postContentDocumentToText,
 } from '@kosmo/core/post-content/server';
 import { eq, ne, sql } from 'drizzle-orm';
-import { createFedifyExecutionContext } from './fedify-execution';
 import { setInboundObservabilityReporter } from './inbound-observability';
 import type { DocumentLoader, InboxContext } from '@fedify/fedify';
 import type * as CoreDb from '@kosmo/core/db';
 import type * as CoreSeed from '@kosmo/core/db/seed';
 import type * as CoreServices from '@kosmo/core/services';
 import type { findPostByActivityPubUri as findPostByActivityPubUriType } from './activitypub-post-uri';
-import type { FedifyExecutionContext } from './fedify-execution';
+import type { FedifyExecutionContext } from './federation';
 import type { handleInboundCreate as handleInboundCreateType } from './inbound-create';
 
 const publicOrigin = 'http://127.0.0.1:4173';
@@ -1156,7 +1155,7 @@ describe('inbound Create dispatch', () => {
         undefined,
         audience,
       ),
-      { contextData: createFedifyExecutionContext() },
+      { contextData: { db } },
     );
     const sharedResponse = await fixture.federation.fetch(
       await fixture.createSignedCreateRequest(
@@ -1166,7 +1165,7 @@ describe('inbound Create dispatch', () => {
         undefined,
         audience,
       ),
-      { contextData: createFedifyExecutionContext() },
+      { contextData: { db } },
     );
 
     assert.equal(personalResponse.status, 202, await personalResponse.text());
@@ -1195,7 +1194,7 @@ describe('inbound Create dispatch', () => {
           objectUri,
           new URL('https://remote.example/activities/create-concurrent-personal'),
         ),
-        { contextData: createFedifyExecutionContext() },
+        { contextData: { db } },
       ),
       fixture.federation.fetch(
         await fixture.createSignedCreateRequest(
@@ -1203,7 +1202,7 @@ describe('inbound Create dispatch', () => {
           objectUri,
           new URL('https://remote.example/activities/create-concurrent-shared'),
         ),
-        { contextData: createFedifyExecutionContext() },
+        { contextData: { db } },
       ),
     ]);
 
@@ -1234,11 +1233,11 @@ describe('inbound Create dispatch', () => {
           new URL('https://remote.example/activities/create-reply-personal'),
           parentUri,
         ),
-        { contextData: createFedifyExecutionContext() },
+        { contextData: { db } },
       ),
       fixture.federation.fetch(
         await fixture.createSignedCreateRequest('/inbox', replyUri, null, parentUri),
-        { contextData: createFedifyExecutionContext() },
+        { contextData: { db } },
       ),
     ]);
