+++ packages/fedify/src/inbound-local-recipient.ts
@@ -2,7 +2,7 @@ import { ActivityPubActors, db, first, Instances, Profiles } from '@kosmo/core/d
 import { InstanceKind, InstanceState, ProfileState } from '@kosmo/core/enums';
 import { and, eq } from 'drizzle-orm';
 import type { InboxContext } from '@fedify/fedify';
-import type { FedifyExecutionContext } from './fedify-execution';
+import type { FedifyExecutionContext } from './federation';
 
 export const resolveInboundLocalRecipient = async (
   context: InboxContext<FedifyExecutionContext>,
