+++ packages/fedify/src/inbound-reject-follow.ts
@@ -13,7 +13,7 @@ import {
 } from './inbound-observability';
 import type { InboxContext } from '@fedify/fedify';
 import type { Follow } from '@fedify/vocab';
-import type { FedifyExecutionContext } from './fedify-execution';
+import type { FedifyExecutionContext } from './federation';
 
 export const handleInboundRejectFollow = async ({
   context,
